# tests/hospital-api-v1_2_openapi.yaml
openapi: 3.1.0
info:
  title: Hospital Management API
  version: 1.2.0
  description: |
    Patients, doctors, departments, OPD schedules, appointments, and the
    hospital information the call centre's voice agent reads to callers.

    This API is the SINGLE SOURCE for everything the voice agent says.
    Structured data (schedules, availability, prices, hours, rooms) is read
    live at the moment a caller asks. Prose (department descriptions,
    services, FAQ articles) is authored here and copied into the agent's
    search index by a sync job that polls `updated_since` (see "Sync").

    ## What changed in 1.2.0

    - NEW `Holidays` — hospital-wide closures. Availability returns a
      `holiday` status and booking refuses the date with `409 HOLIDAY`.
    - NEW `visit_type` on appointments (`new`, `follow_up`,
      `report_showing`) and an optional separate report-showing queue per
      session (`report_capacity`).
    - NEW `POST /schedules/import` — upload the OPD room-allocation
      spreadsheet (or CSV), preview the result, then apply it.
    - NEW `GET /audit-log` — who changed what, when, with before/after.
    - NEW `Hospital` profile — address, hotlines, emails, OPD and desk
      hours (replaces the voice agent's `CRM_OPENING_HOURS` setting).
    - NEW `Packages`, `Services` (including facility rooms) and `Articles`
      (FAQ and general content).
    - CHANGED `Department` — adds `description_bn` and `conditions`.
...

    ## What changed in 1.1.0

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - APIKeyHeader: []
...
paths:
...
  /doctors/{doctor_id}/schedules:
    parameters:
      - $ref: '#/components/parameters/DoctorId'
    get:
      tags: [Schedules]
      summary: List Doctor Schedules
      operationId: list_doctor_schedules
      description: |
        The doctor's weekly OPD sessions. A weekday with no row means NO
        SESSION that day — there is no "unknown" state.
      parameters:
        - name: effective_on
          in: query
          required: false
          schema: { type: [string, 'null'], format: date }
          description: |
            Only rows valid on this date (`valid_from <= date` and
            (`valid_to` is null or `valid_to >= date`)). Omit for all rows,
            past and future.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ScheduleOut' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Schedules]
      summary: Create Doctor Schedule
      operationId: create_doctor_schedule
      description: |
        Requires `admin` scope. Refused with `409 SCHEDULE_OVERLAP` if it
        overlaps another session of the same doctor on the same weekday
        within an overlapping validity range, and `409 ROOM_CONFLICT` if the
        room/desk is already allocated to another doctor for an overlapping
        time on that weekday.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduleCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduleOut' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`SCHEDULE_OVERLAP` or `ROOM_CONFLICT`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /schedules/{schedule_id}:
    parameters:
      - $ref: '#/components/parameters/ScheduleId'
    get:
      tags: [Schedules]
      summary: Get Schedule
      operationId: get_schedule
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduleOut' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Schedules]
      summary: Update Schedule
      operationId: update_schedule
      description: |
        Requires `admin` scope. Takes effect immediately for every future
        date the row covers. To change a rota from a future date, end-date
        this row (`valid_to`) and create a new one with `valid_from` instead.

        Existing appointments are NEVER moved or cancelled by this call. Any
        scheduled appointment from today onward that no longer falls inside
        a session is returned in `affected_appointments` so staff can
        contact those patients.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduleUpdate' }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduleChangeResult' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`SCHEDULE_OVERLAP` or `ROOM_CONFLICT`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Schedules]
      summary: Delete Schedule
      operationId: delete_schedule
      description: |
        Requires `admin` scope. Refused with `409 HAS_FUTURE_APPOINTMENTS`
        while scheduled appointments from today onward are booked into this
        session. To retire a session, set `valid_to` instead.
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`HAS_FUTURE_APPOINTMENTS`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /doctors/{doctor_id}/exceptions:
    parameters:
      - $ref: '#/components/parameters/DoctorId'
    get:
      tags: [Schedules]
      summary: List Schedule Exceptions
      operationId: list_schedule_exceptions
      parameters:
        - name: date_from
          in: query
          required: false
          schema: { type: [string, 'null'], format: date }
          description: Defaults to today.
        - name: date_to
          in: query
          required: false
          schema: { type: [string, 'null'], format: date }
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ScheduleExceptionOut' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Schedules]
      summary: Create Schedule Exception
      operationId: create_schedule_exception
      description: |
        Requires `admin` scope. Leave, a cancelled session, or an extra
        one-off session on a single date. Takes effect immediately.

        As with schedule updates, existing appointments are NOT cancelled:
        those on the date that no longer fall inside a session are returned
        in `affected_appointments`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ScheduleExceptionCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScheduleExceptionResult' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`SCHEDULE_OVERLAP` or `ROOM_CONFLICT` (extra_session only)'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /exceptions/{exception_id}:
    parameters:
      - name: exception_id
        in: path
        required: true
        schema: { type: integer }
    delete:
      tags: [Schedules]
      summary: Delete Schedule Exception
      operationId: delete_schedule_exception
      description: |
        Requires `admin` scope. Refused with `409 HAS_FUTURE_APPOINTMENTS`
        for an `extra_session` that has bookings.
      responses:
        '204': { description: Deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`HAS_FUTURE_APPOINTMENTS`'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ── Availability (NEW, computed) ─────────────────────────────────────────
...
components:
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    DoctorId:
      name: doctor_id
      in: path
      required: true
      schema: { type: integer }
    ScheduleId:
      name: schedule_id
      in: path
      required: true
      schema: { type: integer }
...
  responses:
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HTTPValidationError' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: '`FORBIDDEN_SCOPE` — the API key''s scope does not allow this operation.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    ErrorResponse:
      type: object
      description: NEW in 1.1.0. Body of every business-rule error.
      required: [code, detail]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        detail:
          type: string
          description: Human-readable, English. Not for parsing.
        next_available_date:
          type: [string, 'null']
          format: date
          description: |
            On `NO_SESSION`, `SESSION_FULL` and `HOLIDAY`: the doctor's next
            date (within 31 days) with a free serial, so a client can offer
            it directly.
        holiday_name:
          type: [string, 'null']
          description: NEW in 1.2.0. On `HOLIDAY`.
        candidates:
          type: [array, 'null']
          items: { type: object }
          description: Optional supporting rows (e.g. the clashing schedule on `SCHEDULE_OVERLAP`).
    ErrorCode:
      type: string
      enum:
        - NOT_FOUND
        - PATIENT_NOT_FOUND
        - DOCTOR_NOT_FOUND
        - FORBIDDEN_SCOPE
        - PAST_DATE
        - NO_SESSION
        - SESSION_REQUIRED
        - SESSION_FULL
        - DUPLICATE_BOOKING
        - DOCTOR_NOT_ACCEPTING
        - SCHEDULE_OVERLAP
        - ROOM_CONFLICT
        - HAS_DOCTORS
        - HAS_FUTURE_APPOINTMENTS
        - DUPLICATE_NAME
        - IDEMPOTENCY_KEY_REUSED
        - HOLIDAY
        - DUPLICATE_DATE
        - IMPORT_HAS_ERRORS
        - UNSUPPORTED_FILE
    Weekday:
      type: string
      enum: [saturday, sunday, monday, tuesday, wednesday, thursday, friday]
...
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items: { $ref: '#/components/schemas/ValidationError' }
    ValidationError:
      type: object
      required: [loc, msg, type]
      properties:
        loc:
          type: array
          items:
            anyOf: [{ type: string }, { type: integer }]
        msg: { type: string }
        type: { type: string }

    # ── patients (unchanged shape) ─────────────────────────────────────────
...
    PatientOut:
      type: object
      required: [name, age, mobile_number, id, created_at]
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 150 }
        age: { type: integer, minimum: 0, maximum: 150 }
        mobile_number: { type: string, maxLength: 20 }
        created_at: { type: string, format: date-time }
        updated_at: { type: [string, 'null'], format: date-time }

    # ── departments (NEW) ──────────────────────────────────────────────────
...
    DoctorOut:
      allOf:
        - $ref: '#/components/schemas/DoctorBase'
        - type: object
          required: [id, name, is_active, accepting_appointments, created_at]
          properties:
            id: { type: integer }
            department:
              anyOf:
                - $ref: '#/components/schemas/DepartmentSummary'
                - type: 'null'
            created_at: { type: string, format: date-time }
            updated_at: { type: [string, 'null'], format: date-time }
            deleted_at: { type: [string, 'null'], format: date-time, description: NEW in 1.2.0. }
...
    DepartmentSummary:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        name_bn: { type: [string, 'null'] }

    # ── schedules (NEW) ────────────────────────────────────────────────────
    SessionPlace:
      type: object
      properties:
        room: { type: [string, 'null'], maxLength: 30, description: 'e.g. "3042"' }
        floor: { type: [string, 'null'], maxLength: 30, description: 'e.g. "3rd"' }
        desk: { type: [string, 'null'], maxLength: 10, description: 'e.g. "A"' }
    ScheduleCreate:
      allOf:
        - $ref: '#/components/schemas/SessionPlace'
        - type: object
          required: [weekday, start_time, end_time, max_patients]
          properties:
            weekday: { $ref: '#/components/schemas/Weekday' }
            start_time: { type: string, format: time, description: 'Local, e.g. "09:00"' }
            end_time:
              type: string
              format: time
              description: Must be after `start_time`. Sessions do not cross midnight.
            max_patients:
              type: integer
              minimum: 1
              description: Serials available in this session (its capacity).
            slot_minutes:
              type: [integer, 'null']
              minimum: 1
              description: |
                When set, serial N is given the time `start_time + (N-1) *
                slot_minutes`. When null, every serial shows `start_time`
                and patients are seen in serial order.
            report_capacity:
              type: integer
              minimum: 0
              default: 0
              description: |
                NEW in 1.2.0. Separate serials for `report_showing` visits,
                which then do not use up the regular `max_patients`. 0 means
                report visits take regular serials.
            valid_from:
              type: [string, 'null']
              format: date
              description: First date this row applies. Null means already in effect.
            valid_to:
              type: [string, 'null']
              format: date
              description: Last date this row applies. Null means open-ended.
            is_active: { type: boolean, default: true }
    ScheduleUpdate:
      type: object
      description: Every field optional; omitted fields are unchanged.
      properties:
        weekday:
          anyOf:
            - $ref: '#/components/schemas/Weekday'
            - type: 'null'
        start_time: { type: [string, 'null'], format: time }
        end_time: { type: [string, 'null'], format: time }
        room: { type: [string, 'null'], maxLength: 30 }
        floor: { type: [string, 'null'], maxLength: 30 }
        desk: { type: [string, 'null'], maxLength: 10 }
        max_patients: { type: [integer, 'null'], minimum: 1 }
        slot_minutes: { type: [integer, 'null'], minimum: 1 }
        report_capacity: { type: [integer, 'null'], minimum: 0 }
        valid_from: { type: [string, 'null'], format: date }
        valid_to: { type: [string, 'null'], format: date }
        is_active: { type: [boolean, 'null'] }
    ScheduleOut:
      allOf:
        - $ref: '#/components/schemas/ScheduleCreate'
        - type: object
          required: [id, doctor_id, created_at]
          properties:
            id: { type: integer }
            doctor_id: { type: integer }
            created_at: { type: string, format: date-time }
            updated_at: { type: [string, 'null'], format: date-time }
    ScheduleChangeResult:
      type: object
      required: [schedule, affected_appointments]
      properties:
        schedule: { $ref: '#/components/schemas/ScheduleOut' }
        affected_appointments:
          type: array
          items: { $ref: '#/components/schemas/AppointmentDetailOut' }
          description: |
            Scheduled appointments from today onward that no longer fall
            inside any session of this doctor. They are NOT changed — staff
            must contact the patients and cancel or rebook.
    ExceptionType:
      type: string
      enum: [leave, cancelled, extra_session]
      description: |
        - `leave` — the doctor is away; every session that date is off.
        - `cancelled` — one session that date is off (`schedule_id` required).
        - `extra_session` — a one-off session that date (`start_time`,
          `end_time`, `max_patients` required).
    ScheduleExceptionCreate:
      allOf:
        - $ref: '#/components/schemas/SessionPlace'
        - type: object
          required: [date, type]
          properties:
            date: { type: string, format: date }
            type: { $ref: '#/components/schemas/ExceptionType' }
            schedule_id:
              type: [integer, 'null']
              description: Required for `cancelled`; ignored otherwise.
            start_time: { type: [string, 'null'], format: time }
            end_time: { type: [string, 'null'], format: time }
            max_patients: { type: [integer, 'null'], minimum: 1 }
            slot_minutes: { type: [integer, 'null'], minimum: 1 }
            report_capacity: { type: [integer, 'null'], minimum: 0 }
            note:
              type: [string, 'null']
              maxLength: 500
              description: Staff-internal reason. Omitted from responses for `agent`-scoped keys.
    ScheduleExceptionOut:
      allOf:
        - $ref: '#/components/schemas/ScheduleExceptionCreate'
        - type: object
          required: [id, doctor_id, created_at]
          properties:
            id: { type: integer }
            doctor_id: { type: integer }
            created_at: { type: string, format: date-time }
    ScheduleExceptionResult:
      type: object
      required: [exception, affected_appointments]
      properties:
        exception: { $ref: '#/components/schemas/ScheduleExceptionOut' }
        affected_appointments:
          type: array
          items: { $ref: '#/components/schemas/AppointmentDetailOut' }
          description: Appointments on that date that no longer fall inside a session. NOT changed.

    # ── availability (NEW) ─────────────────────────────────────────────────
...
    AppointmentStatus:
      type: string
      enum: [scheduled, completed, cancelled]
    AppointmentSource:
      type: string
      enum: [voice_agent, call_center, front_desk, web, other]
    VisitType:
      type: string
      description: NEW in 1.2.0.
      enum: [new, follow_up, report_showing]
    QueueName:
      type: string
      description: |
        NEW in 1.2.0. `report` only exists for sessions with
        `report_capacity > 0`.
      enum: [regular, report]
      default: regular
...
    AppointmentOut:
      type: object
      required: [id, patient_id, doctor_id, appointment_date, appointment_time,
                 serial_number, status, created_at]
      properties:
        id: { type: integer }
        patient_id: { type: integer }
        doctor_id: { type: integer }
        appointment_date: { type: string, format: date }
        appointment_time:
          type: string
          format: time
          description: Derived from the session and serial.
        serial_number:
          type: integer
          minimum: 1
          description: NEW in 1.1.0. Position in the session's queue; 1 is first.
        visit_type: { $ref: '#/components/schemas/VisitType' }
        queue:
          $ref: '#/components/schemas/QueueName'
          description: NEW in 1.2.0. `serial_number` counts within this queue.
        schedule_id: { type: [integer, 'null'], description: NEW in 1.1.0. }
        exception_id: { type: [integer, 'null'], description: NEW in 1.1.0. }
        session_start: { type: [string, 'null'], format: time, description: NEW in 1.1.0. }
        session_end: { type: [string, 'null'], format: time, description: NEW in 1.1.0. }
        room: { type: [string, 'null'], description: NEW in 1.1.0. Copied from the session at booking time. }
        floor: { type: [string, 'null'], description: NEW in 1.1.0. }
        desk: { type: [string, 'null'], description: NEW in 1.1.0. }
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        source:
          anyOf:
            - $ref: '#/components/schemas/AppointmentSource'
            - type: 'null'
        cancel_reason: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        created_at: { type: string, format: date-time }
        updated_at: { type: [string, 'null'], format: date-time }
    AppointmentDetailOut:
      allOf:
        - $ref: '#/components/schemas/AppointmentOut'
        - type: object
          required: [patient, doctor]
          properties:
            patient: { $ref: '#/components/schemas/PatientOut' }
            doctor: { $ref: '#/components/schemas/DoctorOut' }

    # ── holidays (NEW in 1.2.0) ────────────────────────────────────────────
...